remove the duplication between`Board` and `SudokuBoard` (#7)
Migrate all uses of `SudokuBoard` to just use a modified version
of`Board`. This removes duplication between the two which exists because
of bootstrapping this crate from the C++ version in pieces.
`SudokuBoard` is now just a type alias for `Board` (plus the relevant
Sudoku-related type parameters).

The following other changes were made:

- Introduce the `GameDefinition` trait to model the "zones" defined for
grid games instead of the `ZoneMetadata`.
- Add `GenericGameDefinition` as the implementation of `GameDefinition`
which most grid games should use. Use it to implement
`SudokuGameDefinition`.
- Refactor `Board` so that the applicable `GameDefintion` and the
board's capacity are type parameters.
- Introduce `Board::to_puzzle_str` and `Board::from_puzzle_str` for
converting to/from puzzle strings.